fix api category

+++ src/models/category.model.js
@@ -1,4 +1,5 @@
 const { pool } = require("../config/db");
+const { get } = require("../routes/category.routes");
 
 const CategoryModel = {
   async getAll() {
@@ -13,28 +14,33 @@ const CategoryModel = {
     return res.rows[0];
   },
 
-  async create({ key, label, color, japanese, kana, romaji }) {
+  async create({ key, label, color, japanese, kana, romaji, display_order }) {
     const res = await pool.query(
-      `INSERT INTO category (key, label, color, japanese, kana, romaji)
-       VALUES ($1, $2, $3, $4, $5, $6) RETURNING *`,
-      [key, label, color, japanese, kana, romaji]
+      `INSERT INTO category (key, label, color, japanese, kana, romaji, display_order)
+       VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING *`,
+      [key, label, color, japanese, kana, romaji, display_order]
     );
     return res.rows[0];
   },
 
-  async update(key, { label, color, japanese, kana, romaji }) {
+  async update(key, { label, color, japanese, kana, romaji, display_order }) {
     const res = await pool.query(
       `UPDATE category
-       SET label=$1, color=$2, japanese=$3, kana=$4, romaji=$5
-       WHERE key=$6 RETURNING *`,
-      [label, color, japanese, kana, romaji, key]
+       SET label=$1, color=$2, japanese=$3, kana=$4, romaji=$5, display_order=$6
+       WHERE key=$7 RETURNING *`,
+      [label, color, japanese, kana, romaji, display_order, key]
     );
     return res.rows[0];
   },
 
   async delete(key) {
-    await pool.query("DELETE FROM category WHERE key=$1", [key]);
-    return true;
+    if (( await this.getByKey(key).length) === 0 || await this.getByKey(key) == null || await this.getByKey(key) == undefined) {
+      console.log("Category not found");
+      return false;
+    } else {
+      await pool.query("DELETE FROM category WHERE key=$1", [key]);
+      return true;
+    }
   },
 };
